+++ test/main.cpp
@@ -11,9 +11,9 @@
 #include "src/base/num.hpp"
 #include "src/base/str.hpp"
 #include "src/base/utf8.hpp"
-#include "src/base/iter.hpp"
 #include "src/base/regex.hpp"
 #include "src/base/vector.hpp"
+#include "src/base/iterator.hpp"
 
 #include "src/data/ini.hpp"
 #include "src/data/json.hpp"